PhaserBase in lib-builtin-effects
